Change metrics value type in float

+++ plugin/models/FrameRateManager.cs
@@ -12,8 +12,8 @@
             framesCounter++;
         }
 
-        public override int CalculateMetric() {
-            int frameRate = (int) (framesCounter / timer);
+        public override float CalculateMetric() {
+            float frameRate = (framesCounter / timer);
 
             if(timer - updateFrequency < updateFrequency) {
                 framesCounter = 0;
